split alloc defs
